[Context] Now saving each uniform variable to a lookup table
Also moved all gfx context related stuff to their own folder

+++ Source/Runtime/context/opengl/opengl_vao.cpp
@@ -3,7 +3,7 @@
 
 #include <glad/glad.h>
 
-#include "renderer/context/vao.h"
+#include "context/vao.h"
 
 namespace polos
 {
@@ -13,12 +13,12 @@ namespace polos
         int32 alignment { GL_NONE };
         glGetIntegerv(GL_UNIFORM_BUFFER_OFFSET_ALIGNMENT, &alignment);
     
-        auto const vrt_size { static_cast<int32>(vertices.size_bytes()) };
-        auto const ind_size { static_cast<int32>(indices.size_bytes())  };
+        auto const vrt_size { vertices.size_bytes() };
+        auto const ind_size { indices.size_bytes()  };
         
-        auto const align = [](int32 const length, int32 const alignment) -> int64 {
-            const int32  misalignment { length & ( alignment - 1 ) };
-            const int32  padding      { (alignment - misalignment) & (alignment - 1) };
+        auto const align = [](int64 const length, int64 const alignment) -> int64 {
+            const int64  misalignment { length & ( alignment - 1 ) };
+            const int64  padding      { (alignment - misalignment) & (alignment - 1) };
             return length + padding;
         };
